requirements.txt
streamlit
pyyaml
streamlit-authenticator
pandas
plotly